sort findings by severity when inserting to report.xml
Sorting the findings on initial insertion does at least sort them by severity when copying the first time. That already eases the manual cleanup in that file. A more complete implementation should sort the findings after any modification of severity though, because with this implementation later changes do not have an effect.