raise exception when adding to non-existant section
When attempting (non-)findings to report.xml the target section might not exist. Previously the failed attempt resulted in a warning, which may disappear in the (otherwise successful) build logs as warning. Instead an exception causes the report conversion to crash, but only if the report should in fact include an item in the target section